requirements.txt
# ══════════════════════════════════════════════════════════════════
#  AlphaGrid v6 — Python dependencies
#  Tested on macOS M4 / Python 3.11 & 3.12
#  Install: pip install -r requirements.txt
#  PyTorch is installed SEPARATELY (see SETUP.md — needs MPS flags)
# ══════════════════════════════════════════════════════════════════

# ── Core science ──────────────────────────────────────────────────
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.13.0

# ── Market data (free, no API key required) ───────────────────────
yfinance>=0.2.50

# ── API server ────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
python-multipart>=0.0.9

# ── Database ──────────────────────────────────────────────────────
sqlalchemy>=2.0.29

# ── Auth (JWT + bcrypt) ───────────────────────────────────────────
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# ── ML — gradient boosting ────────────────────────────────────────
lightgbm>=4.3.0
scikit-learn>=1.4.0
imbalanced-learn>=0.12.0

# ── Technical analysis ────────────────────────────────────────────
pandas-ta>=0.3.14b

# ── Async / HTTP ──────────────────────────────────────────────────
aiohttp>=3.9.0
httpx>=0.27.0
requests>=2.31.0
websockets>=12.0

# ── Utilities ─────────────────────────────────────────────────────
loguru>=0.7.2
pyyaml>=6.0.1
python-dotenv>=1.0.1
python-dateutil>=2.9.0
pytz>=2024.1

# ─────────────────────────────────────────────────────────────────
# OPTIONAL — uncomment what you need
# ─────────────────────────────────────────────────────────────────

# Alpaca broker (paper + live equities)
# alpaca-trade-api>=3.0.2

# OANDA broker (forex)
# oandapyV20>=0.7.2

# Robinhood (unofficial — use at own risk)
# robin-stocks>=2.2.0
# pyotp>=2.9.0

# News sentiment (FinBERT — downloads ~500MB model on first use)
# transformers>=4.40.0
# feedparser>=6.0.10

# Reddit scraping (WSB, r/stocks)
# praw>=7.7.1

# Numba JIT (makes indicators 10–50× faster, optional)
# numba>=0.59.0

# Hyperparameter tuning
# optuna>=3.6.0

# Statistical models (OU process, GARCH)
# statsmodels>=0.14.0
